Fixed 32-bit iOS compile error in Replication_Tests.m
Pasin and I didn't see this because the error only occurs when building
32-bit, and the iOS unit tests target was set to build the active
architecture only in debug builds. I fixed it to build all archs, so
we'll see this sort of error in the future.

+++ Unit-Tests/Replication_Tests.m
@@ -123,7 +123,7 @@
 - (void) cookiesChanged: (NSNotification*)n {
     CBLCookieStorage* storage = n.object;
     _changedCookies = storage.cookies;
-    Log(@"%@ changed: %lu cookies", storage, _changedCookies.count);
+    Log(@"%@ changed: %lu cookies", storage, (unsigned long)_changedCookies.count);
 }
